requirements.txt
aiohttp
fastapi
uvicorn
pytest-asyncio
